sih-recommendation-engine-prototype-1/requirements.txt
streamlit
faiss-cpu
numpy
requests
python-dotenv
google-generativeai
sentence_transformers